requirements.txt
# Core ML/NLP Libraries
sentence-transformers
transformers
torch
spacy>=3.0.0,<4.0.0
# Note: After installing spacy, download a language model, e.g.:
# python -m spacy download hu_core_news_lg

# Vector Search
faiss-cpu # or faiss-gpu if you have CUDA installed and configured

# Graph Database (Example: Neo4j)
neo4j

# Data Handling
numpy
pandas

# Reinforcement Learning
gym

# Web UI
Flask # Replaced streamlit with Flask

# Utilities
tqdm

# Optional Document Loaders (Install if you need to process .docx or .rtf)
# python-docx
# striprtf

faiss-cpu # or faiss-gpu if you have CUDA installed and configured
# Note: If you want to use the Neo4j graph database, install the Neo4j Python driver